Modifications to make a deployable script.
I needed to make several changes to allow this script to be sent via the
SQL service we use to push updates. I've created a python script that
will build a single script to wrap all the outputted SQL in a function,
execute it, and prevent any exceptions from occurring, such as if
logging is already deployed. It should silently exit if that is the
case.