Added new functionality, and a new pruning func
logging.prune_size takes a maximum table size parameter. The function
does its best to guess how many records it needs to prune to meet the
requirement of the function. Before preceding, a CLUSTER is performed to
remove table bloat so the size calculation is accurate. This CLUSTER is
performed on the logging_xx_query_id index.

logging.modified now also checks for a temporary table named
logging.logging_disable (which does not need any columns). If this table
is present, the changes that would be logged by logging.modified are not
recorded. This behavior may change to something more efficient in the
future.

Disabled automatic installation of a PGAgent job for pruning. This
should be handled instead on a case-by-case basis, after inspecting
available disk space for logs.